Reparação da Estrutura de Ficheiros
As funcionalidades de Autorização encontram-se agora no AuthController
