+++ SmartMobileWebApp/backend/config/main.php
@@ -45,7 +45,7 @@ return [
             'enablePrettyUrl' => true,
             'showScriptName' => false,
             'rules' => [
-                ['class' => 'yii\rest\UrlRule','controller' => 'api/user'],
+                ['class' => 'yii\rest\UrlRule','controller' => 'api/auth'],
 
             ],
         ],
